[Fix] #90 로직 최적화 및 오류 수정
